test(React): test Blog component rendering
