+++ app/javascript/packs/Blog/index.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Blog = (props) => {
+  const { title } = props;
+  return (
+    <div>
+      Blog
+      {title}
+      !
+    </div>
+  );
+};
+
+Blog.defaultProps = {
+  title: 'Infinite Scrolling',
+};
+
+Blog.propTypes = {
+  title: PropTypes.string,
+};
+
+export default Blog;
