+++ app/javascript/packs/index.jsx
@@ -1,25 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import PropTypes from 'prop-types';
-
-const Blog = (props) => {
-  const { title } = props;
-  return (
-    <div>
-      Blog
-      {title}
-      !
-    </div>
-  );
-};
-
-Blog.defaultProps = {
-  title: 'Infinite Scrolling',
-};
-
-Blog.propTypes = {
-  title: PropTypes.string,
-};
+import Blog from './Blog';
 
 document.addEventListener('DOMContentLoaded', () => {
   ReactDOM.render(
